app/requirements.txt
asgiref==3.4.1
Brotli==1.0.9
cffi==1.15.0
cryptography==36.0.0
cssselect2==0.4.1
distlib==0.3.4
Django==3.2.9
django-cors-headers==3.10.1
django-environ==0.8.1
django-filter==21.1
django-rest-knox==4.1.0
djangorestframework==3.12.4
filelock==3.4.1
fonttools==4.27.1
html5lib==1.1
importlib-metadata==4.8.3
importlib-resources==5.4.0
Markdown==3.3.6
mysqlclient==1.4.6
Pillow==8.4.0
platformdirs==2.4.0
pycparser==2.21
pydyf==0.1.2
pyphen==0.11.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
typing_extensions==4.0.1
virtualenv==20.13.0
weasyprint==54.1
webencodings==0.5.1
XlsxWriter==3.0.2
zipp==3.6.0
zopfli==0.1.9
